Use move_to_data_array for timeseries logs
